Deleted the old MaxHeap class and renamed the MaxHeapTest to MaxHeap.
Tested the new class, and it works as expected.

Added some print statements to the MainClass file to let the user know what is going on.
