+++ DataStructuresAndSorts/Assignment9/MaxHeap.cs
@@ -4,45 +4,19 @@ using System.Collections.Generic;
 
 namespace Assignment9
 {
-    class MaxHeap<T> : IEnumerable<T> where T : IComparable
+    class MaxHeap<T> : Heap<T> where T : IComparable
     {
-        private T[] UnderlyingArray;
-        private int ElementCount = 0;
-        private int OldCount = 0;
-        private bool IsHeap = true;
-
+        public MaxHeap()
+        {
+            UnderlyingArray = new T[5];
+        }
+        
         public MaxHeap(int starting_size)
         {
             UnderlyingArray = new T[starting_size];
         }
 
-        public MaxHeap()
-        {
-            UnderlyingArray = new T[5];
-        }
-
-        public int Count
-        {
-            get
-            {
-                return ElementCount;
-            }
-        }
-
-        public void Clear()
-        {
-            UnderlyingArray = new T[UnderlyingArray.Length];
-        }
-
-        public T this[int index]
-        {
-            get
-            {
-                return UnderlyingArray[index];
-            }
-        }
-
-        public void Add(T new_value)
+        public override void Add(T new_value)
         {
             if (!IsHeap)
                 BuildHeap();
@@ -55,7 +29,7 @@ namespace Assignment9
             ElementCount++;
         }
 
-        public T PopTop()
+        public override T PopTop()
         {
             if (!IsHeap)
                 BuildHeap();
@@ -66,11 +40,11 @@ namespace Assignment9
             UnderlyingArray[ElementCount - 1] = default(T);
             ElementCount--;
             Sink();
-            
+
             return max;
         }
 
-        public void Sort()
+        public override void Sort()
         {
             if (!IsHeap)
                 BuildHeap();
@@ -88,21 +62,42 @@ namespace Assignment9
             ElementCount = OldCount;
         }
 
-        public void BuildHeap()
+        private void Sink()
         {
-            if (IsHeap)
-                return;
+            int current_index = 0;
 
-            int last_item = OldCount - 1;
-
-            for (int index = 0; index < OldCount / 2; index++)
+            while (current_index < ElementCount)
             {
-                Swap(index, last_item);
-                last_item--;
-            }
+                int max_child = current_index;
+                int left_child = FindLeft(current_index);
+                int right_child = FindRight(current_index);
 
-            ElementCount = OldCount;
-            IsHeap = true;
+                if (left_child >= ElementCount)
+                    break;
+
+                if (right_child >= ElementCount)
+                    max_child = left_child;
+
+                if (right_child < ElementCount &&
+                    UnderlyingArray[left_child].CompareTo(UnderlyingArray[right_child]) == 0)
+                    max_child = left_child;
+
+                if (right_child < ElementCount &&
+                    UnderlyingArray[left_child].CompareTo(UnderlyingArray[right_child]) > 0)
+                    max_child = left_child;
+
+                if (right_child < ElementCount &&
+                    UnderlyingArray[left_child].CompareTo(UnderlyingArray[right_child]) < 0)
+                    max_child = right_child;
+
+                if (UnderlyingArray[max_child].CompareTo(UnderlyingArray[current_index]) == 0)
+                    break;
+
+                if (UnderlyingArray[max_child].CompareTo(UnderlyingArray[current_index]) > 0)
+                    Swap(current_index, max_child);
+
+                current_index = max_child;
+            }
         }
 
         private void Swim()
@@ -121,91 +116,6 @@ namespace Assignment9
 
                 current_index = parent_index;
             }
-        }
-
-        private void Swap(int first, int second)
-        {
-            T temp = UnderlyingArray[first];
-            UnderlyingArray[first] = UnderlyingArray[second];
-            UnderlyingArray[second] = temp;
-        }
-
-        private void Sink()
-        {
-            int current_index = 0;
-
-            while (current_index < ElementCount)
-            {
-                int max_child_index = current_index;
-                int left_child_index = FindLeft(current_index);
-                int right_child_index = FindRight(current_index);
-
-                if (left_child_index >= ElementCount)
-                    break;
-
-                if (right_child_index >= ElementCount)
-                    max_child_index = left_child_index;
-
-                if (right_child_index < ElementCount &&
-                    UnderlyingArray[left_child_index].CompareTo(UnderlyingArray[right_child_index]) == 0)
-                    max_child_index = left_child_index;
-
-                if (right_child_index < ElementCount && 
-                    UnderlyingArray[left_child_index].CompareTo(UnderlyingArray[right_child_index]) > 0)
-                    max_child_index = left_child_index;
-
-                if (right_child_index < ElementCount && 
-                    UnderlyingArray[left_child_index].CompareTo(UnderlyingArray[right_child_index]) < 0)
-                    max_child_index = right_child_index;
-
-                if (UnderlyingArray[max_child_index].CompareTo(UnderlyingArray[current_index]) == 0)
-                    break;
-
-                if (UnderlyingArray[max_child_index].CompareTo(UnderlyingArray[current_index]) > 0)
-                    Swap(current_index, max_child_index);
-
-                current_index = max_child_index;
-            }
-        }
-
-        private int FindParent(int current_index)
-        {
-            return (current_index - 1) / 2;
-        }
-
-        private int FindLeft(int current_index)
-        {
-            return (current_index * 2) + 1;
-        }
-
-        private int FindRight(int current_index)
-        {
-            return (current_index * 2) + 2;
-        }
-
-        private void Resize()
-        {
-            T[] new_array = new T[UnderlyingArray.Length * 2];
-
-            for (int index = 0; index < UnderlyingArray.Length; index++)
-                new_array[index] = UnderlyingArray[index];
-
-            UnderlyingArray = new_array;
-        }
-
-        public IEnumerator<T> GetEnumerator()
-        {
-            for (int index = 0; index < ElementCount; index++)
-            {
-                T temp = UnderlyingArray[index];
-                //Console.WriteLine("Current Value: {0}", temp);
-                yield return temp;
-            }
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
         }
     }
 }
